Use file based stream to reduce Apache POI memory foot print.
