+++ rented-properties/src/main/java/org/egov/cpt/service/IReadExcelService.java
@@ -1,0 +1,9 @@
+package org.egov.cpt.service;
+
+import java.io.File;
+
+import org.egov.cpt.models.RentDemandResponse;
+
+public interface IReadExcelService {
+    public RentDemandResponse getDatafromExcel(File file, int sheetIndex);
+}
