+++ rented-properties/src/main/java/org/egov/cpt/web/controllers/ReadExcelController.java
@@ -1,17 +1,18 @@
 package org.egov.cpt.web.controllers;
 
+import java.io.File;
+import java.net.URI;
 import javax.validation.Valid;
 
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang.StringUtils;
 import org.egov.cpt.models.ExcelSearchCriteria;
 import org.egov.cpt.models.RentDemandResponse;
 import org.egov.cpt.models.RequestInfoWrapper;
-import org.egov.cpt.service.ReadExcelNewFormatService;
 import org.egov.cpt.service.ReadExcelService;
 import org.egov.cpt.util.FileStoreUtils;
 import org.egov.tracer.model.CustomException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.UrlResource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -29,32 +30,26 @@ public class ReadExcelController {
 
 	private ReadExcelService readExcelService;
 	private FileStoreUtils fileStoreUtils;
-	private ReadExcelNewFormatService readExcelNewFormatService;
 
 	@Autowired
-	public ReadExcelController(ReadExcelService readExcelService, FileStoreUtils fileStoreUtils,
-			ReadExcelNewFormatService readExcelNewFormatService) {
-		this.readExcelService = readExcelService;
+	public ReadExcelController(ReadExcelService readExcelService, FileStoreUtils fileStoreUtils) {
 		this.fileStoreUtils = fileStoreUtils;
-		this.readExcelNewFormatService = readExcelNewFormatService;
+		this.readExcelService = readExcelService;
 	}
 
 	@PostMapping("/read")
 	public ResponseEntity<RentDemandResponse> readExcel(@Valid @RequestBody RequestInfoWrapper requestInfoWrapper,
 			@Valid @ModelAttribute ExcelSearchCriteria searchCriteria) {
-		log.info("Start controller method readExcel() Request:" + searchCriteria);
 		try {
+			log.info("Start controller method readExcel() Request:" + searchCriteria);
 			String filePath = fileStoreUtils.fetchFileStoreUrl(searchCriteria);
-			RentDemandResponse data = new RentDemandResponse();
-			if (StringUtils.isNotBlank(filePath)) {
-				Integer formatFlag = readExcelService.checkFormatOfexcel(new UrlResource(filePath).getInputStream(), 0);
-				if (1 == formatFlag) {
-					data = readExcelNewFormatService.getDatafromExcel(new UrlResource(filePath).getInputStream(), 0);
-				} else if (0 == formatFlag) {
-					data = readExcelService.getDatafromExcel(new UrlResource(filePath).getInputStream(), 0);
-				}
-				log.info("End controller method readExcel formatFlag :" + formatFlag);
+			if (StringUtils.isBlank(filePath)) {
+				throw new CustomException("FILE_NOT_FOUND", "Cannot find rent history file that is uploaded");
 			}
+			File tempFile = File.createTempFile("File" + System.currentTimeMillis(), ".xlsx");
+			FileUtils.copyURLToFile(new URI(filePath).toURL(), tempFile);
+			RentDemandResponse data = this.readExcelService.getDatafromExcel(tempFile, 0);
+			tempFile.delete();
 			log.info("End controller method readExcel Demand data:" + data.getDemand().size() + " & Payment data:"
 					+ data.getPayment().size());
 			return new ResponseEntity<>(data, HttpStatus.OK);
