+++ rented-properties/src/test/java/org/egov/cpt/service/ExcelStatementTests.java
@@ -27,6 +27,7 @@ public class ExcelStatementTests {
         this.utils = new RentCollectionUtils();
     }
 
+    @SuppressWarnings("deprecation")
     private List<RentAccountStatement> _testSheetWithIndex(String excelFileToParse, int index) {
         InputStream inputStream = ReadExcelServiceTests.class.getClassLoader().getResourceAsStream(excelFileToParse);
         RentDemandResponse rentDemandResponse = this.readExcelService.getDatafromExcel(inputStream, index);
